added some methods, attributes
the code, now, only includes one scheduled method. And all the other methods are added to the corresponding class.

the angle of the ball bounce now depends on the position where it hits the bat. however, there is still a need for the speed of the ball to depend on the speed of the bat.